Change function output.

+++ nautilus_trader/core/functions.pxd
@@ -11,7 +11,7 @@ from cpython.datetime cimport datetime
 
 cpdef double fast_round(double value, int precision)
 cpdef double basis_points_as_percentage(double basis_points)
-cdef int get_obj_size(obj)
+cdef long get_obj_size(obj)
 cpdef str format_bytes(double size)
 cpdef str pad_string(str string, int length, str pad=*)
 cpdef str format_zulu_datetime(datetime dt)
